requirements.txt
albumentations<2.0
matplotlib
numpy
opencv-python
opencv-python-headless
pandas
Pillow
PyQt5
PyQt5-sip
PyYAML
QtPy
scikit-learn
torch
torchvision
tqdm
